+++ lib/arachni/processes/instances.rb
@@ -146,6 +146,7 @@ class Instances
 
         instance = connect( info['url'], info['token'] )
         instance.framework.set_as_master
+        instance.opts.grid_mode = :aggregate
         instance
     end