+++ lib/arachni/rpc/server/framework/distributor.rb
@@ -337,7 +337,7 @@ module Distributor
     def cleaned_up_opts
         opts = @opts.to_h.deep_clone.symbolize_keys
 
-        (%w(spawns rpc_socket grid dir rpc_port rpc_address pipe_id neighbour pool_size) |
+        (%w(spawns rpc_socket grid_mode dir rpc_port rpc_address pipe_id neighbour pool_size) |
             %w(lsmod lsrep rpc_instance_port_range load_profile delta_time) |
             %w(start_datetime finish_datetime)).each do |k|
             opts.delete k.to_sym